Add statistics collector and measure time in nanoseconds

+++ src/main/java/org/semanticweb/elk/justifications/experiments/Record.java
@@ -2,10 +2,10 @@ package org.semanticweb.elk.justifications.experiments;
 
 public class Record {
 	
-	public final long time;
+	public final double time;
 	public final int nJust;
 	
-	public Record(final long time, final int nJust) {
+	public Record(final double time, final int nJust) {
 		this.time = time;
 		this.nJust = nJust;
 	}
